Separate SimpleMap definitions to a different package
