+++ easynav_simple_maps_manager/include/easynav_simple_maps_manager/SimpleMapsManager.hpp
@@ -35,7 +35,7 @@
 
 #include "easynav_core/MapsManagerBase.hpp"
 #include "easynav_common/types/MapTypeBase.hpp"
-#include "easynav_simple_maps_manager/SimpleMap.hpp"
+#include "easynav_simple_common/SimpleMap.hpp"
 
 namespace easynav
 {
