+++ easynav_simple_maps_manager/tests/CMakeLists.txt
@@ -1,9 +1,9 @@
-ament_add_gtest(simple_maps_tests simple_maps_tests.cpp)
-target_include_directories(simple_maps_tests PUBLIC
+ament_add_gtest(simple_mapsmanager_tests simple_mapsmanager_tests.cpp)
+target_include_directories(simple_mapsmanager_tests PUBLIC
   $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
   $<INSTALL_INTERFACE:include>
 )
-target_link_libraries(simple_maps_tests
+target_link_libraries(simple_mapsmanager_tests
   ${PROJECT_NAME}
 )
-ament_target_dependencies(simple_maps_tests ${dependencies})
+ament_target_dependencies(simple_mapsmanager_tests ${dependencies})
